Refactored code: no God class <levelhandler> -> added level 1 and 0 their rightful classes; Stringbuilder in levelZero introduced
